Refactor Total part to be an actual record in DB
Since GLASS questions have to be able to refer individually to any of all parts
we need to have an actual record to be refered to. This means the asumption
of "an absence of part, means the total" is no longer true. Instead we can
use the read-only Part::isTotal() property to find out the one and single part
representing the total.

Also parts are now pre-existing in database (via migrations), this allow GLASS
to be independent of JMP import process.