Revert commit 170d049, do not sum children if answer exists
We must not sum children if an answer exists for a filter. The presence
of an answer override everything that may or may not exists in children.

What commit 170d049 tried to fix must actually be fixed by modifying our
filter tree structure and/or entering answers on the right filters. So
basically fix the original data, not our computing algorithm.