Clean up Assertion and expose new 'permissions' property for all objects
Assertions can now be used to assert that our model has a specific state
in addition to role and permission given to user.

REST API now respond to ?fields=permissions by returning an array of
actions and whether it is allowed. This should be used on client side
to enable/disable buttons accordingly