Change the way double onblur event fire is managed.
Angular ng-blur effect fire twice event. To avoid to save modifications 2 times, we had to wait until
previous save has end.

The second fired is null. Now this property is used to detect the second fired event and avoir double save.
Not more needs to wait the end of the previous saving.