Improved and Shared are based on ratios available from all questionnaires #2126
It was previously based on the current questionnaire ratio, which made it
available for only a very few years, making the trend line completely wrong.
By using the average of all ratios, we are much closer to correct result
with a trend line available for all years.

But the sum of trend lines are not equals to "Improved + shared", because
those are trend lines, and not the values at questionnaire level. So it's
slightly different than Excel files, but has the advantage to keep logic
consistent with existing computing (no special cases required).