Can import several questionnaires for a single survey #2126
We used to assume that imported surveys had only one questionnaire. This
is wrong for Russia (RLMS96). This commit allow to import several questionnaires
for a same survey that appears several time in Excel files.

It also means that from now on, we cannot re-import data to overwrite them,
because it will always create new questionnaires.