Computation now support several questionnaires for one survey #2126
We used to use survey code as key element to identify questionnaire,
which made impossible to have several questionnaires for a single
survey on a chart. We now use questionnaire ID as key to avoid confusion
and be more robust.

Also NumericJson now use more robust regexp to supporting quotes