Correct value for total part of regression lines #2288
We used the sum of used questionnaire population to compute total part
from complementary parts. This was wrong and the much simpler way to
do it is to use the population of the year being computed. It is not
related at all with questionnaire anymore. This give us exact results
on chart for Water, total part.