Prevent cyclic reference in formulas #2294
If cyclic reference happen in formulas, the second occurence of the
formula will be entirely ignored, falling back on other formulas if
any, or on raw result.

Amongst other things it solve the case of Azerbaijan Estimation based
on total part which is itself based on urban and rural part, which are
themselves based on the Estimation.