Implemented a template for e-mails. Change target of the notification "questionnaire validated" .
Validated questionnaire notification is now sentto Role editor instead of questionnaire creator.
(Mark Haltmeier has them all created and nobody will be notified with first approach).

The e-mail template come from Mailchimp and need to be compiled with this tool :
http://beaker.mailchimp.com/inline-css.

The php tags <? ?> and -> are replaces by htmlentities and need to be replaced when
the main template is changed. email_src.phtml is the uncompiled and email.phtml is compiled.

A better solution homemade should be done to avoid compiling or implement Mailchimp api to compile on automaticly with build.sh.