Prevent over-extended trend line for Total part #2289
Trend line for total part could be extented way to far in the past
because we in case of (correct) NULL result we defaulted on non-complementary
computation. We basically changed the computation method in the middle of
the trendline: usually complementary, but sometimes not.

We now strictly use complementary computation for total part, without any
fallback. It is both simpler and more correct, but may introduce invalid
result for countries without with only data for Total part.