100% correct values for Bangladesh #2289
Improved and Shared filter are now defined as the ratio itself, as
seen in original coutry files. This slightly counter-intertuitive
since a filter is used for a ratio, but it allow us to have the
exact same behavior as Excel files, resulting in 100% exact data
for all Bangladesh.

In order to achieve this, we introduced a new syntax to be able to
get a list of all available filter values for all questionnaires.