Support referencing Filter value on level 1 or 2 #2359
Trinidad and Tobago have formulas for filters which refers to
filter value of total part on level 1 (Tables_W, not GraphData_W).
This was previously impossible to modelize, because computation was
entirely in either level 1 or 2. By specifying which level we refer
to in the formula, we have a more expressive, yet more flexible engine.
Allowing us to "change" level during computation.