Replace hardcoded complementary total with formula #2360
We used to hardcode the fact that high filters results for total part
was the sum of other parts. However, this is not always the case. By moving
this kind of behavior to Rule's formulas, we simplify our codebase while
giving enough flexibility to fullfill all use cases.

In particular this allow us to get correct results for Afghanistan.s