Cache for getValuePercent missed hit it shouldn't #2358
If asking for a questionnaire data, but another questionnaire had no
answer at all, then the cache was not filled for the second questionnaire.
This lead to an extra SQL query when asking data for the second
questionnaire, whereas we should already have it in cache.

We now ensure that we get all questionnaire for a Geoname in cache, even
empty questionnaire. This avoid a few queries but does not improve speed
significantly.