Revert "Open select2 when in tag mode and get focus via keyboard #2550"
This reverts commit 3c885b654c1536477fe5d3f6aed943524fb92c56, because
it introduce cumbersome behavior. When navigating with only TAB key
select2 in tag mode would select the first element in addition to
moving to next form field. Also using the mouse to select an element
would immediately re-open the select2 making a "blinking" effect.

Since the original "problem" was not critical to begin with, we agreed
to do nothing at all instead of breaking even more the GUI.