Adjustator now gives up faster in edge cases #2671
If the reference filter is composed of two child filters and the first child
is already higher than the target, but we are modyfing the second child,
that means that whatever we do we cannot reach the target. The best we can do
is override the second child to exactly 0, and accept a less than ideal final result.

In this edge case we used to try very hard to reach 0 for the second child, but never
reached it (because of how the heuristic works). Now, however, we are able to "jump"
quickly to 0, thus avoiding precious computing time.