Share calculator within ChartController #2671
Interestingly enough, sharing the calculator within
ChartController does not improve significantly the response
time. In most cases there is almost not difference at all.

And in the most extreme edge case, when plotting all
filterSet at once, for total part, on a single chart, we save
only 30% of time (from ~9s to ~6s). But since it doesn't
break anything, we keep the optimization anyway.