Apply complementary logic on questionnaires #2769
We already used complementary computing for trend lines, but did not
do it on questionnaires. This lead to confusing chart with points
(questionnaires) with a very different value than trend lines.

The difference was, in part, when a ratio was applied to "Total improved",
then the ratio was not taken into account to compute "Other improved",
which was wrong. By using complementary computing, we automatically
benefit from all ratios or special cases done for "Total improved".