More accurate way to compute Shared and Improved #2769
We used to compute Shared and Improved in a very particular way where
they did not represent a population coverage but a ratio. The ratio
was then used for trend lines. It gaves correct trend lines, but the
questionnaires (points) on chart were very confusing.

We now use something much closer to original Excel files, where both
Shared and Improved have an actual population coverage for all
questionnaires. Those values are partly based on the regression value
of "Improved + shared" and the ratio mentionned before. This is
counter-intuitive because we use regression value to compute
non-regression value, but it is correct. This is the reason why we
merged both calculator classes into a single one, because we must be
able to get regression value even when computing non-regression value.

One more trick is that Shared and Improved are both based on the
average of all available ratios of all questionnaires. And not on the
ratio of the current questionnaire. That's unusual, and that means
that all questionnaires always have a value for Shared and Improved
independently of whether there is a ratio or not for that particular
questionnaire.

That also mean that rules for Shared and Improved will need to be
maintained by end-user whenever a new ratio is available.