Distinct dash styles for countries #2760
Dash styles are used to differentiate countries. Colors are used
to differentiate filters. If a special serie is drawn (such as ignored
elements or projection), then original series are "transparent" and
dotted, independently of countries.

Moreover when a projection is active all filters are computed with
the projection taken into account. This is consistent with when we
ignore elements. And it also allow to see wether there are side-effects
on other filters.

Since all selected filters are drawn when projecting, we do not have
to force-inject filters (eg: their ancestors). It is now up to the
end-user to select whatever filters he want to see on chart. This is
technically simpler, faster and more flexible for end-user. It also avoids
to "pollute" the chart with things that may not be wanted and could not
be removed.