Projection lines are correctly displayed #2795
In case of projection we always refresh all lines (both original and adjusted).
This is slighlty sub-optimal, but since original lines will always change
dash style, we must refresh them from server. The only case were lines would
not change and could be kept is if already have a projection, and change a
parameter. In that case original lines would remain identical and does not need
to be re-computed from server. However this is considered an edge-case not worth
the effort (for now).