Aggregator show only questionnaire for a single geoname #2956
We do not show questionnaire recursively for all geonames anymore because
it does not make much sense to begin with. And in the case of the world that
would be 2000+ points on chart per filter, which is way too much data to be
displayed at once. It was absolutely unreadable. Removinf this recursivity
speed up the process and make it everything easier to use.