Differences after projection were not shown
Difference data were only attached to a single adjusted serie.
But if by chance that particular serie was ignored because identical
to the non-adjusted variant, then difference were not shown at all.

We now attach difference information on all adjusted series, and in
addition we simplify client side because there can only be a single
table of difference at a time.