Cache warm-up mixed up results #3035
This was due because we compute all filters in a single pass, meaning
re-computing several time the same filter, but in slightly different
context. We forgot to include used formula in that context, which ended
up with mixing all kind of wrong result. That resulting in gigantic
differences for end-results. Eg: Tokelau, Total Improved, Rural was
off by 100%.