Could not properly edit alternate names #3113
Question name was absent from loaded data, so we always assumed we had
to save as official name instead of alternate.

In addition to that our API returned an empty array instead of an empty
object which confused our javascripts. We now use a dummy value to
avoid having empty things at all.