Alternate name input never show official name #3113
It used to work, but was confusing for user. When trying to delete
the alternate name, the official would come up, giving the impression
that either nothing happened, or something was mixed up.

Recently it was broken and even more confusing. The official name would
only appear on after an update.

To remove all ambiguity, save CPU cycle and simplify our code, we limit
this input to strictly show and manage alternate names.