Avoid wrongly returning 0 if value is NULL #3406
When IF condition is TRUE, but the value to be returned is NULL,
then PHPExcel will return 0. PHPExcel behavior is consistent with
official Excel documentation (though it's a really weird thing to do...).
So in GIMS we must be extra careful when using IF that the returning
value in case of TRUE condition must be a number, otherwise we'll end
up with wrong results.