Migrate from PostgreSQL to MariaDB 10.1
While PostgreSQL still is technically preferable to MariaDB, the maintenance
effort required for the infrastructure is not worth it anymore. Especially
MariaDB 10 drastically improved over MySQL 5 that was first considered back
when we chose PostgreSQL. However there are still a few downside to this migration

Eight `CHECK` were dropped and should be re-introduced when moving to MariaDB 10.2
that will support them. Meanwhile that means that there is a risk of data corruption.
But the risk is deemed minor, because the app is almost not used at all, especially
not to edit things.

MariDB does not support triggers that modify the same table that the trigger is on.
That means that the recursive deletion mechanism of dependent rule when deleting a
rule had to be moved on PHP side, and thus it is not recursive anymore, but only
delete the first 'level' of dependency. Again there is a risk of data corruption,
but we at least treat the most common cases, and rule deletion should not happen
very often (if ever), especially with a multi-level dependency. This will likely
not be fixable in the forseeable future, unless we improve the PHP code (probably
moving all triggers to PHP and improving to support recursiveness).

Some API results were not yet ordered deterministically, resulting in some changes
in expected test data.

Finally geonames geometry were somehow never imported correctly in PostgreSQL
(all identical values, hence they were all set to NULL values, since they are
never used anywhere in the application anyway.